+++ backend/molgenis-emx2-fairmapper/src/main/java/org/molgenis/emx2/fairmapper/transform/RdfTransformer.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.fairmapper;
+package org.molgenis.emx2.fairmapper.transform;
 
 import org.eclipse.rdf4j.repository.sail.SailRepository;
 import org.molgenis.emx2.io.tablestore.TableStore;
