+++ backend/molgenis-emx2-fairmapper/src/test/java/org/molgenis/emx2/fairmapper/transform/SparqlSelectRdfTransformerTest.java
@@ -1,4 +1,4 @@
-package org.molgenis.emx2.fairmapper;
+package org.molgenis.emx2.fairmapper.transform;
 
 import static org.eclipse.rdf4j.model.util.Statements.statement;
 import static org.eclipse.rdf4j.model.util.Values.iri;
