+++ backend/molgenis-emx2-fairmapper/src/test/resources/org/molgenis/emx2/fairmapper/transform/query.sparql
